collectionScript/requirements.txt
psycopg2
paho-mqtt
python-dotenv
